+++ src/BIP44CoinTypeNode.test.ts
@@ -98,7 +98,7 @@ describe('BIP44CoinTypeNode', () => {
       for (const input of inputs) {
         await expect(
           BIP44CoinTypeNode.fromJSON(input as any, arbitraryCoinType),
-        ).rejects.toThrow('Invalid key: Must be a non-zero 32-byte key.');
+        ).rejects.toThrow('Invalid value: Must be a non-zero 32-byte buffer.');
       }
 
       await expect(
